backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-genai
python-dotenv==1.0.0
werkzeug==3.0.1
sqlalchemy>=2.0.36
flask-jwt-extended==4.6.0
bcrypt==4.1.2
